部分已有配表的自动化配置，设置，教学系统搭建以及部分优化

+++ Assets/PersonaCards/Cards/Hands/HandEvaluationResult.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using PersonaCards.Core;
 
 namespace PersonaCards.Cards.Hands
 {
@@ -29,7 +30,7 @@ namespace PersonaCards.Cards.Hands
 
         public int BaseChips => Definition.BaseChips;
 
-        public int BaseMultiplier => Definition.BaseMultiplier;
+        public decimal BaseMultiplier => Definition.BaseMultiplier;
 
         public IReadOnlyList<string> ScoringCardIds => _scoringCardIds;
     }
